PythonRobotFramework/py/requirements.txt
async-generator==1.10
attrs==21.2.0
behave==1.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
h11==0.12.0
idna==3.3
outcome==1.1.0
parse==1.19.0
parse-type==0.5.2
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==21.0.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
splinter==0.16.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
